updates program publication check test, ditch obsolete py publication check test.
the program acceptance test was actually in the program-year directory,
and the program-year acceptance test was in the program directory.
weird.
i switched them around before deleting the py-test.
